Added dice-ext.scm
The extended dice gets rid of the unneccesarry do-loop when only one parameter
is given to the script. Additionally, a third parameter was introduced to
modify the die roll by the given real number, with a cond-statement added to
check whether the third parameter (y) is zero, positive or negative and to
print the results accordingly.