Modified dice-ext.scm
The original (if)-function has been replaced by a (cond)-function in order to
print usage information instead of an error because of an empty argument.

Additionally, the printing of error messages in the case of non-numerical
input for the second and third parameter has been improved in order to give
more specific instructions.