stklos: modified > dice-ext.scm
Refactored the complete main-function, getting rid of "error handling",
instead printing the usage if the amount arguments is <1 or >3, letting stklos
handle the error messages (albeit slightly more cryptic) for non-numbers as
input. Additionally, the refactored code uses a second macro (roll) to deal
with rolling the die and actually formatting the result, giving a leaner and
more easily readable (main) body.

Note that while there are guile-implementations possible, this version makes
heavy use of stklos-specific fundamentals, namely (argv) and (argc) for
command-line argument handling.

guile: added > dice-guile.scm

This is a guile implementation of the dice-program as implemented above for
stklos, using the "~@d" format option which automatically sets the correct
algebraic sign. Unlike the stklos-version above, this version cannot take the
shortcut via (argv) and (argc) for command-line handling and thus has to use
regular list functions for a slightly more cluttered (main) function, most of
which however is mitigated by using the (vref) macro here as well.