stklos: modified > schemeblog.scm
This revision introduces the first changes in order to make (tag) work on
arbitrary output ports. The current default for (main) is output to a
string-port, but the overall change in the macro makes it fairly trivial to
write directly to files (without redirecting the shell's stdout).

changed: (tag) uses "port" instead of "#t"
changed: indentation (buffer)
added: (make) as an attempt of a main function with content stored in a list
changed: (main) now wraps the called function to print the outport-port