racket: added > eve-crest.rkt
This script scrapes the CCP's official CREST API sov campaign end-point in
order to display a quick overview of sovereignty structures that are currently
reinforced.

racket: added > eve-rischwa.rkt

This script scrapes rischwa.net's API for existing coalitions and prints them
in a format that can be easily imported into a spreadsheet.