racket: modified > eve-intel.rkt
changed: use a generic function to build api-query-strings from a list of
groupIDs instead of hardcoding them, which makes this function infinitely
extensible for an arbitrary amount of groupIDs. This can also trivially be
rewritten in the future to use typeIDs instead of groupIDs, which would allow
us to pull each type of super's data individually, or it would allow this tool
to function as a general api disection tool that can filter arbitrary typeIDs.

changed: make api-pulls conditional, so they only run when we are not running
interactively, to avoid unnecessary data downloading/handling.