racket: modified > eve-intel.rkt
changed: use a function instead of a variable assignment for the main api
call and remove the check for interactive running, since we already check for
interactive mode in the main function and thus only need to call the api call
at all if we are running non-interactively.

reverted: print times again regardless of interactive status. These can always
be filtered out via awk, sed etc., but this allows us to write a thin wrapper
around the script to pull APIs or chain API pulls while still running
interactively