racket: modified > eve-intel_towers.rkt
Changed the script so that we can now run it interactively, for a list of
regions given on the command-line (-r/--regions), with a given start date
(-d/--date) and end date (-e/--end-date) for the overall query, instead of the
hardcoded values we had before. All regions get appended and then converted to
csv as final output.