racket: modified > eve-intel_towers.rkt
Modified script so that it now takes arbitrary groupIDs via a command-line
argument (-g/--group), so it is no longer restricted to towers. A later commit
will reflect this change by also changing the script name. The script also has
toggles for the display killmail links, a list of attackers and to switch
between the groupid as attackers or victims. Some internal functions have been
renamed accordingly.