racket: added > eve-api_check.rkt
This script takes the polled zkill super data as a source and looks up
the current corporation & alliance for each listed pilot from the EVE
XML APIv2. In order to do so, it first looks up each pilot's
characterID, which is then used in a subsequent query for the pilots'
current affiliation.