racket: modified > lib/main.rkt
changed: replace json-api/xml-api with macros that allow us to choose
between using gzip and using plain transfer, as well as an option to
choose sensible defaults. Current defaults are gzip for json (used for
zkill & CREST) and plain for xml-api, since CCP doesn't seem to manage
to make the APIv2 connection not die otherwise.