racket: modified > eve-digest.rkt
added: use locationID (similar to moonID) via command-line switch in
order to display where the killmail occured (useful for POCOs etc.), but
retain moonID for a simplified moon display. We optionally save the
locationID in the place of the moonID in raw output mode, so it might be
worth renaming the appropriate field in the table at some point.