racket: modified > lib/eve-string_tools.rkt
fixed: regexp for moon itemName translation for both split-moon-display
and simplify-moon-display, since it failed for names containing a suffix
in parentheses (e.g. Uplingur IV (Ndoria) - Moon 16)

fixed: simplify conditional by allowing a second word (e.g. "Prime") to
be optional, which means we do not need to duplicate the whole regexp
operation twice for cases with and without a second word