racket: added > src/eve-corporation_clean-duplicates.rkt
This file allows us to delete corporation entries in customCorporations
where more than one corporation uses the same corporation ticker. Since
this is only possible after a former corporation has been closed, we can
simply grab a list of all duplicate tickers, keep the last one (i.e. the
highest corporationID) and delete the others.

It is perhaps advisable to backup customCorporations before running this
script, since it touches (and deletes) actual data.

The whole purpose of this operation is to afterwards introduce a new
constraint for the table, marking corporation tickers as additional
unique keys alongside the primary key (corporationID), so that we can
make sure we don't grab multiple corporations with the same ticker when
we do tower scans.