racket: modified > servlets/dashboard.rkt
changed: modify the list of available regions in the region filter to
only show regions which exist in the relevant database table, which
allows the user to see straight away for which regions we have any
data (or, in the case of timers, events) at all. "Empty" regions can
still be called directly via URL parameters (e.g. "URL?region=XYZ") in
case the user would like to see (for whatever reason) an empty dataset
for a specific region.