racket: added > servlets/dashboard/citadel-database.rkt
This page is roughly the citadel equivalent to moon-database and
goo-database, though with one crucial difference: citadel-database
allows users to delete citadels from the underlying database through the
page itself, allowing for more control, but also adding a significant
risk. Removed citadels are therefore retained in a separate table to
make sure we do not lose painstakingly collected data to a malicious
user. Since we are mixing our traditional GET-based URL parameters for
region selection and POST-based delete requests, we also need an
intermediary function that deals with deleting data before redirecting
us back to the url contained in the request (which conveniently includes
URL parameters).