racket: modified > servlets/dashboard/citadel-database.rkt
added: check the request headers for its current referer; if the referer
matches recon.tendollarbond.com, strip the leading "/recon" from the
redirect url, pass the url as-is otherwise. This is necessary to work
around a quirk in the remote proxy setup when it comes to generation of
url paths.