racket: added > src/eve-clean_dscans.rkt
This script can be used to delete obsolete dscans from a direction given
as prefix (-p <dir>|--prefix <dir>). The script checks which dscans are
referenced in either our moon or citadel database tables and excludes
those ids from the purge list. Similarly, we only clean (by default)
dscans that are both unreferenced AND older than 60 days. This script
has a "quiet" output that only raises errors when a file cannot be
deleted (permissions etc.), which might be useful for running via
cron. Otherwise, a short status message and the number of purged files
is printed.