racket: modified > lib/eve-html_tools.rkt
This commit adds the option to pass an "audience" (one of:
any|recon|recon-l) to (create-html-navigation), and more links for more
"privileged" audiences (i.e. base -> some -> all). By default, we only
show a link to Dashboard/dscan. In order to not break compatibility, we
alternatively allow to specify a list of links directly, provided we do
not simultaneously pass an audience.

added: css rule to switch navigation wrap to column when viewport <680px
added: option to pass an "audience" to (create-html-navigation)