racket: added > src/eve-api_characters.rkt
This script is used to update affiliations for characters from local
scans that we already hold in an SQL table, working in the same way as
eve-api_check.rkt, which is used to update super pilots. We are keeping
the same looping mechanism and limits in order to handle potential
virtual memory pressure issues.