racket: modified > servlets/dashboard/parse.rkt
This commit adds the option to paste local scans into the dscan tool,
which will then get parsed and output similarly to regular dscans. To do
so, we slightly rename the existing workflow, but otherwise follow the
same steps. Unlike dscans, we need to do some additional lifting for
local scans before saving them to disk in order to be able to save a
fully parsed version containing a list of struct:sql-character to disk,
which lets us avoid reparsing or overwriting affiliations that were
current at the time of scanning.

added: condition for parsing local scans
added: sentinel that decides whether input is a dscan or local scan
changed: rename exec-parse to exec-prepare-dscan
added: function to write local scan to disk, update sql & redirect
added: function to prepare local scan for output before sending it on