racket: modified > lib/eve-auth_jwt.rkt
This commit reorganises the JWT auth library and, more importantly, adds
a function (try-authorization-header) that can be used within a function
call to create a html-navigation as #:audience in order to set
navigation links depending on the presence of an "Authorization" HTTP
header and depending on its contents of its JWT subject. The same
function can also be used to try and extract a username from the same
header if provided.

added: parameter to set jwt-secret globally
added: parameter to set jwt-audiences globally
added: parameter to set jwt-issuer globally
changed: use global parameters for (create-token) & (verify-token)
added: function to try and extract a subject (or username) from auth