racket: added > src/eve-task_dispatcher.rkt
This script implements a rudimentary task dispatcher that is intended to
replace the separate invocation of background updaters via crontab with
one monolithic dispatcher that uses Racket's thread system to handle
various API polls internally, which means we do not have to load a full
racket environment for each script invocation, as would be the case for
crontab. Hardcoded intervals are provided for now, as is rudimentary
exception "handling" (discard and continue). We are running two threads
by default.