racket: modified > lib/eve-sql_supers.rkt
This commit replaces the intelSuperLatest view with a table (as a
pseudomaterialized view), adding a lot more fields to each entry in the
process (IDs + Names, constellation, Tickers etc.). This new table
serves as the basis for a new intelSuperWatchlist and for changes to
eve-sql_canary.rkt, which will be made in a future commit.

added: pseudomaterialized table intelSuperLatestMV
changed: base intelSuperWatchlist on the new table
added: insert/delete triggers for intelSuperLatestMV
added: function to create triggers