racket: modified > servlets/dashboard/result.rkt
This commit adds provisions to check submitted corporation tickers
against ESI to make sure the given corporation actually exists. If it
does, continue as usual, if it does not, prompt the user to enter
a *valid* corporation ticker (checking again). This also integrates this
check into both moon-scan-result and citadel-scan-result to avoid
parsing (and writing) data with invalid corporation tickers. The overall
implementation follows that of citadel-scan-result's location-try
system. If the corporation is valid *AND* we don't have it in our own
database, parse and insert it while we are it already. This commit
changes esi-fill-alliance to use this result directly, though we will
not make use of this option at this moment.

Caveat: Following these changes, it would be possible to use
corporation-try also as input for esi-fill-alliance. However, we are
currently passing corporation directly from POST, so while we
technically have a superfluous API look within esi-fill-alliance, this
isn't really a big deal.

added: form to submit a corrected corporation ticker entry
changed: if->cond for the final html, since we have >2 options
added: corporation-try check via esi-try-corporation
added: resolve unknown (to us) but valid corporations
changed: include corporation-try in moon-scan-result
changed: include corporation-try in citadel-scan-result

racket: modified > lib/eve-esi_tools.rkt

added: handle number inputs for esi-fill-alliance directly (see above)